Added option to leave docker installed on headnode

+++ install.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
 
+OPTIND=1
+
+docker_allow=0 #default to NOT installing docker; must be 0 or 1
+
+while getopts ":d" opt; do
+  case ${opt} in
+    d) docker_allow=1
+      ;;
+    \?) echo "BAD OPTION! $opt TRY AGAIN"
+      exit 1
+      ;;
+  esac
+done
+
 if [[ ! -e /etc/slurm/openrc.sh ]]; then
   echo "NO OPENRC FOUND! CREATE ONE, AND TRY AGAIN!"
   exit 1
@@ -26,9 +40,12 @@ dnf -y install http://repos.openhpc.community/OpenHPC/2/CentOS_8/x86_64/ohpc-rel
        centos-release-openstack-train
 
 dnf config-manager --set-enabled powertools
-dnf config-manager --set-disabled docker-ce-stable
 
-dnf -y remove containerd.io.x86_64 docker-ce.x86_64 docker-ce-cli.x86_64 docker-ce-rootless-extras.x86_64
+if [[ ${docker_allow} == 1 ]]; then
+  dnf config-manager --set-disabled docker-ce-stable
+  
+  dnf -y remove containerd.io.x86_64 docker-ce.x86_64 docker-ce-cli.x86_64 docker-ce-rootless-extras.x86_64
+fi
 
 dnf -y install \
         ohpc-slurm-server \
